+++ xoa/core/__init__.py
@@ -1,0 +1,1 @@
+"""Core low-level routines accelerated with numba"""